for meso

+++ tools/tmp.py
@@ -8,7 +8,7 @@ from scipy.stats import pearsonr
 
 def calculate_glcm(volume, distances=[1], angles=[0, np.pi/4, np.pi/2, 3*np.pi/4]):
     """Calcule les matrices de co-occurrence pour un volume 3D."""
-    glcm = graycomatrix(volume.astype(np.uint8), distances=distances, angles=angles, symmetric=True, normed=True)
+    glcm = graycomatrix(volume, distances=distances, angles=angles, symmetric=True, normed=True)
     return glcm
